[move-prover] Make vnext Boogie backend the default
Removes the old version. The vnext backend is the one which added support for multiple array theories and was off by default to not disrupt release 1.2.

While the vnext is slightly faster as benchmarks in tools/data show, it also adds three new timeouts for transaction scripts. These are followed up on later.

This also adds a new flag `--z3-trace=<function name>` which narrows verification to the given function and produces a file `<function-name>.z3log` which can be fed into jupyterlab.

Apart of this there are no logical code changes in this PR, just boilerplate from removing old `boogie-backend` and renaming `boogie-backend-v2` to this name.

Closes: #8129
