+++ language/diem-framework/modules/TreasuryComplianceScripts.move
@@ -263,6 +263,9 @@ module TreasuryComplianceScripts {
         /// **Access Control:**
         /// Only the account with a Preburn resource or PreburnQueue resource can preburn [[H4]][PERMISSION].
         aborts_if !(exists<Diem::Preburn<Token>>(account_addr) || exists<Diem::PreburnQueue<Token>>(account_addr));
+
+        /// TODO(timeout): this currently times out
+        pragma verify = false;
     }
 
     /// # Summary
