+++ language/diem-framework/releases/artifacts/current/docs/modules/PaymentScripts.md
@@ -179,6 +179,13 @@ Parent VASPs, and Child VASPs can hold balances [[D1]][ROLE][[D2]][ROLE][[D3]][R
 </code></pre>
 
 
+TODO(timeout): this currently times out
+
+
+<pre><code><b>pragma</b> verify = <b>false</b>;
+</code></pre>
+
+
 
 </details>
