+++ language/diem-framework/releases/artifacts/current/docs/modules/TreasuryComplianceScripts.md
@@ -499,6 +499,13 @@ Only the account with a Preburn resource or PreburnQueue resource can preburn [[
 </code></pre>
 
 
+TODO(timeout): this currently times out
+
+
+<pre><code><b>pragma</b> verify = <b>false</b>;
+</code></pre>
+
+
 
 </details>
