+++ language/move-prover/Cargo.toml
@@ -9,7 +9,6 @@ license = "Apache-2.0"
 [dependencies]
 # diem dependencies
 boogie-backend = { path = "boogie-backend" }
-boogie-backend-v2 = { path = "boogie-backend-v2" }
 move-lang = { path = "../move-lang" }
 move-model = { path = "../move-model" }
 docgen = { path = "docgen" }
@@ -28,7 +27,6 @@ clap = "2.33.3"
 codespan = "0.8.0"
 codespan-reporting = "0.8.0"
 futures = "0.3.12"
-handlebars = "3.5.3"
 hex = "0.4.3"
 itertools = "0.10.0"
 log = { version = "0.4.14", features = ["serde"] }
