+++ language/move-prover/boogie-backend/src/prelude/vector-array-intern-theory.bpl
