+++ language/move-prover/boogie-backend/src/prelude/vector-array-theory.bpl
