+++ language/move-prover/boogie-backend/src/prelude/vector-smt-array-ext-theory.bpl
