+++ language/move-prover/boogie-backend/src/prelude/vector-smt-array-theory.bpl
