+++ language/move-prover/boogie-backend/src/prelude/vector-smt-seq-theory.bpl
