+++ language/move-prover/tests/sources/functional/loops.exp
@@ -80,8 +80,8 @@ error: induction case of the loop invariant does not hold
      =     at tests/sources/functional/loops.move:188: loop_with_two_back_edges_incorrect
      =     at tests/sources/functional/loops.move:189: loop_with_two_back_edges_incorrect
      =     at tests/sources/functional/loops.move:191: loop_with_two_back_edges_incorrect
-     =         x = <redacted>
-     =     at tests/sources/functional/loops.move:197: loop_with_two_back_edges_incorrect
+     =         y = <redacted>
+     =     at tests/sources/functional/loops.move:193: loop_with_two_back_edges_incorrect
      =     at tests/sources/functional/loops.move:189: loop_with_two_back_edges_incorrect
 
 error: base case of the loop invariant does not hold
